backend/requirements.txt
fastapi>=0.110.0
uvicorn[standard]>=0.29.0
python-dotenv>=1.0.0
python-multipart>=0.0.9
pydantic>=2.6.0

# PDF extraction
PyMuPDF>=1.24.0

# Vector store
chromadb>=0.5.0
sentence-transformers>=2.6.0

# LLM & agents
crewai>=0.80.0
litellm>=1.30.0
langchain-groq>=0.1.0
